Add initial ACL checks to message consumption and production
This commit uses the previously implemented ACL store to check each inbound request to
consume and produce messages to ensure the client is allowed. The tlsinfo package has
once again been swapped back to clientinfo. The clientinfo package now focuses on being
an abstraction for obtaining client information from various sources via a new Extractor
type that is passed into the unary and stream server interceptors.

The package currently contains two methods of extracting client information, via TLS
certificate details (TLSExtractor) or by gRPC metadata (MetadataExtractor).

The server has been modified to pick an extractor type based on the presence of a TLS
key and certificate on startup. The client has been modified to allow the configuration
of a client identifier that is added to the outbound metadata.

Signed-off-by: David Bond <davidsbond93@gmail.com>