Implement client-side load-balancing and leader discovery
This commit adds the node gRPC service implementation that returns whether or not a node
is the leader in the cluster. It also includes modifications to the client package to handle
multiple node addresses so that leader-only operations are sent to the correct node.

This is done fairly simply, and there are probably caveats with it at this point, but it's a simple
first start. On creation, the client will determine a leader. After each request, if a codes.FailedPrecondition
is returned from a leader-only operation, each connection will be checked again to find the leader,
then the operation is attempted again.

This initial implementation can probably lead to an infinite loop, but works for now.

Signed-off-by: David Bond <davidsbond93@gmail.com>