Remove metadata checks from TLSExtractor
This commit modifies the clientinfo TLSExtractor to no longer check the metadata field to ensure it matches,
the assumption here is that workloads may not always know what their own spiffe identifier is, so we shouldn't
make it a requirement for the workload to know it. It also provides nothing to have that extra check in, as the
spiffe identifier should always be preferred when using TLS.

Documentation has also been added for the client configuration regarding TLS and spiffe. The generate-certs target
has also been updated to use an extension file that embeds a spiffe id in the certificate for the server to read. I've
done some testing and found that this works as intended.

Signed-off-by: David Bond <davidsbond93@gmail.com>