Implement signing gRPC service
This commit adds the gRPC signing service which is used to generate public/private key pairs on demand
for clients. Each unique client may have a single public/private key pair to sign message keys with. An
error is returned if key generation is requested for an existing client.

Included in this commit are changes to the client and e2e tests for the signing service. It does not
include the handling of signed messages from producers.

Signed-off-by: David Bond <davidsbond93@gmail.com>