Implement message key signing when produced
This commit modifies the message gRPC service so that it verifies the message key signature
when a message is produced. The client has been modified to include the key signature as
part of the Sender, rather than as request metadata, this means we can drop having to base64
encode them all the time.

e2e tests have been updated to produce a signed message, consume it and verify it.

Signed-off-by: David Bond <davidsbond93@gmail.com>