Improvements after trying to run cluster in Kubernetes
This commit contains fixes to the kubernetes manifests for deploying a cluster and adds a
check when setting up serf that we do not attempt to join ourselves. This can cause an issue
where the service cannot route to the pod because the pod is not yet ready and listening.

Signed-off-by: David Bond <davidsbond93@gmail.com>