Updated memberlist logger to infer levels with timestamp
This commit changes the logger configuration for the memberlist logger so that it
infers levels with a timestamp. The original configuration did not take timestamps
into account which caused the levels in the memberlist logger to not be inferred.

Memberlist seems to use hclog but its API only exposes a log.Logger, so hclog has
this helper function presumably made for it.

Signed-off-by: David Bond <davidsbond93@gmail.com>