Expose peers in node service
This commit modifies the node gRPC service so that the Describe call returns an array
of all node's advertise addresses. To enable this, each serf peer has been changed
to set its node name to the configured advertise address, when added to the raft
configuration, their advertise address will become the server id, which is returned
in the gRPC call.

Signed-off-by: David Bond <davidsbond93@gmail.com>