Use advertise address as raft local ID
This commit modifies the raft bootstrapping process to use the node's
advertise address as the raft local id. This is required as the
initial implementation in a kubernetes cluster causes the following
response when requesting peers:

```
peers = ["0.0.0.0:5000", "arrebato-1.arrebato.svc.cluster.local, ...]
```

The first element in the array should also be the k8s service address, which
is taken from the advertise address.

Signed-off-by: David Bond <davidsbond93@gmail.com>