Use AppliedIndex instead of LastIndex for FSM check
This commit modifies the FSM to use AppliedIndex rather than LastIndex. This is done
because we want to check the last known index applied to the FSM, rather than the last
known index within the raft log. It was observed during testing with a 6 node cluster that
hundreds of messages sent to the leader would not necesserily make their way to the follow
nodes, this is likely the culprit.

Signed-off-by: David Bond <davidsbond93@gmail.com>