Manually track last applied log index
This commit adds functionality to the FSM to manually track the last applied log index. This is
so that when the raft log is replayed we do not duplicate messages. We can't completely rely on
the raft's AppliedIndex function as it may typically lag behind or not be accurate, so we can
manually store and check it when handling an inbound log.

Signed-off-by: David Bond <davidsbond93@gmail.com>