Make servers leaving/rejoining work better
This commit makes modifications to the general serf+raft node discovery. Firstly, it modifies
the serf member join handling code to remove any servers first that share the identifier of
the new serf member.

Secondly, it sets up serf to use a snapshot file that should be persisted so that on start, the
previously disconnected node can attempt to rejoin the cluster.

Signed-off-by: David Bond <davidsbond93@gmail.com>