First steps in message partitioning
This commit contains the first steps in adding message partitioning to the
application. When creating topics, a partition count is now included that
must be at least one.

When messages are produced, a crc32 hashing function and a modulo operation
are used to generate a uint32 within the range of 0 to the number of partitions.
This is calculated once by the leader node and propagated to the followers by
the raft log.

Future commits will include storing messages in independant streams of partitions
and some mechanism for consuming those, still TBC.

Signed-off-by: David Bond <davidsbond93@gmail.com>