Add node commands triggered by serf membership
This commit adds new commands to the raft log that indicate nodes joining and leaving
the cluster. This is required for the next step of message partitioning where topics
can be allocated to individual nodes on creation and reallocated when a node leaves
the cluster.

To do this, we need some way to persist the state of all nodes within the state storage,
so when the leader finds a node has left or joined it will add a command to the raft
log that all nodes can use to update their state.

Signed-off-by: David Bond <davidsbond93@gmail.com>