Implement node data persistence
This commit builds off the previous one by adding handlers that update the node
storage when the add and remove commands are played via the raft log. These records
will eventually be used to determine which node is allocated which topic.

I've neglected to mention previously that the intention is to allocate topics
to individual nodes that will be the place consumers go to read from those
topics. This should allow us to have proper partitioning with consumer groups
as on a single node we can keep track of which consumer in the group has which
partitions and rebalance them accordingly when members of the consumer group
leave or are added.

Signed-off-by: David Bond <davidsbond93@gmail.com>