Add topic allocation on creation (#38)
This commit modifies the topic and node stacks to allocate topics to a node when they
are created. When a topic is created, the leader looks up the node which has the least
number of topics allocated to it and allocates it there. This should ensure an even
distribution of topics amongst nodes.

In future, we'll need to reallocate topics when a node leaves the cluster and ensure
all topics are allocated when a node obtains leadership.

Signed-off-by: David Bond <davidsbond93@gmail.com>