Add endpoint for listing public signing keys (#42)
This commit adds a new RPC to the signing service that lists all keys stored within
the server. It also updates the CLI to provide a command for listing public keys. Among
the CLI changes is the new table package which provides a nice way to display tables
of information.

Signed-off-by: David Bond <davidsbond93@gmail.com>