Allow nodes to join cluster as non-voters
This commit modifies the raft/serf configuration to allow servers to join the
cluster as non-voters. This essentially makes them never able to become a leader
and therefore always be a read-only replica.

Server's raft port, advertise address and role are also now included in their
serf tags. This should allow the server to filter out any clients that join
the gossip protocol without advertising themselves as a server.

Signed-off-by: David Bond <davidsbond93@gmail.com>