Add configuration for serf-level encryption (#48)
This commit allows the operator of a cluster to enable encryption at the serf layer
using an encryption key. This should prevent snooping of the gossip protocol providing
the key isn't exposed.

The encryption key should be stored as a file that contains 16, 24 or 32 bytes which will
select AES-128, AES-192 or AES-256. A make command has been added to generate a sample key
of 32 bytes using /dev/urandom.

Since we now have 2 separate serf files, one for the cluster state and one for the keyring, they
are now placed in their own serf directory within the configurable data path.

Signed-off-by: David Bond <davidsbond93@gmail.com>