Add persistence logic for topic assignment
This commit adds the raft log handling logic for topic assignment. When the command
comes in via the raft log, the node data in the state is updated to include the name
of the newly assigned topic. All other records are also checked and the topic
is removed from those records.

Signed-off-by: David Bond <davidsbond93@gmail.com>