Ensure node record is created on leadership changes
This commit modifies the leader node to always attempt to create its record
via the raft log when it becomes the leader.

Because we can't guarantee leadership as soon as serf begins finding servers, the
first node's record may not be created. To get around this, we check for leadership
changes and, if we are the leader, attempt to create the node record.

Signed-off-by: David Bond <davidsbond93@gmail.com>