Require clients to consume topics from their assigned node
This commit modifies the client and message service to only allow consuming messages
from the node that the topic is assigned to. This will make distributing messages across
consumers in a group simpler, as we can guarantee that all the consumers for a topic will
be using the same node.

Signed-off-by: David Bond <davidsbond93@gmail.com>