Assign topics to specific nodes and require clients to use those nodes (#93)
* Define raft commands for node state

This commit adds two commands to the node package that occur when nodes are either
added or removed from the cluster. These will eventually be stored in state and used
for topic distribution.

* Add handler and store for persisting node state

This commit adds the command handling and bolt persistence logic for node state. When commands
are written to the raft log they are now handled and node data will be stored.

* Update node state when members are added or leave

This commit modifies the serf handling logic to execute the new add/remove node commands
when new serf members are added. This should keep the state of nodes within the store
up-to-date so we can assign and reassign topics.

* Add topics field to node type

This commit modifies the node protobuf definition to include the list of topics assigned
to the node. This will be returned by the node service so clients know which node to use
for a given topic.

* Add command to indicate topic assignment

This commit adds the AssignTopic command which will be used to indicate that a topic
has been assigned to a specific node.

* Add persistence logic for topic assignment

This commit adds the raft log handling logic for topic assignment. When the command
comes in via the raft log, the node data in the state is updated to include the name
of the newly assigned topic. All other records are also checked and the topic
is removed from those records.

* Assign topic to a node on creation

This commit modifies the topic gRPC handling code to assign the topic to the node
with the least number of assigned topics.

* Ensure node record is created on leadership changes

This commit modifies the leader node to always attempt to create its record
via the raft log when it becomes the leader.

Because we can't guarantee leadership as soon as serf begins finding servers, the
first node's record may not be created. To get around this, we check for leadership
changes and, if we are the leader, attempt to create the node record.

* Expose available topics via node service

This commit modifies the node gRPC service to obtain peer and topic data from
the internal store rather than the raft configuration. Clients will be able to
use this to know which node to go to for each topic.

* Ensure all topics are assigned to a node

This commit modifies the raft and self logic to ensure that all topics
are assigned to a node when either a leadership change occurs or a node
leaves the cluster.

* Sort nodes, topics & keys when listing them

This commit adds sorting to the node, topic & key listing client functions. I noticed in some
cases the list commands were giving non-deterministic output.

* Use sorting to determine best node placement

This commit modifies the node picking logic to use sort.Slice, this seems to yield
a more balanced distribution than the original approach.

* Require clients to consume topics from their assigned node

This commit modifies the client and message service to only allow consuming messages
from the node that the topic is assigned to. This will make distributing messages across
consumers in a group simpler, as we can guarantee that all the consumers for a topic will
be using the same node.

* Add documentation explaining how topics work

This commit adds some basic documentation outlining how topics work and
how they are assigned to nodes. It also includes a small fix for the
kubernetes manifests as the metric port was incorrect.

Signed-off-by: David Bond <davidsbond93@gmail.com>