Add distributed tracing (#100)
This commit adds support for distributed tracing via the OpenTelemetry
collector. It modifies the source code to set up a trace provider and
adds traces to the gRPC, command & handler levels.

The docker-compose has been updated to include the OpenTelemetry collector
and an instance of Jaeger to view the traces. Documentation has also been
added for working with the docker-compose file.

Server configuration flags have been simplified to remove knobs that don't
really need to be configurable.

Signed-off-by: David Bond <davidsbond93@gmail.com>